backend/worker/requirements.txt
psycopg2-binary
redis
SQLAlchemy
python-dotenv
